Move method to UI side
